// old/frontend/src/app/interfaces/language.interface.ts
export interface LanguageTranslations {
  chooseLanguage: string;
  polish: string;
  english: string;
  home1: string;
  home2: string;
  address: string;
  search: string;
  second1: string;
  second2: string;
  step1_1: string;
  step1_2: string;
  step2_1: string;
  step2_2: string;
  step3_1: string;
  step3_2: string;
  third1: string;
  third2: string;
  why1_1: string;
  why1_2: string;
  why2_1: string;
  why2_2: string;
  why3_1: string;
  why3_2: string;
  why4_1: string;
  why4_2: string;
  myAccount: string;
  dontHaveAccount: string;
  createNow: string;
  login: string;
  profile: string;
  orders: string;
  favourites: string;
  opinions: string;
  help: string;
  changeLanguage: string;
  language: string;
  name: string;
  email: string;
  password: string;
  confirmPassword: string;
  registerAccept: string;
  haveAccount: string;
  createAccount: string;
  emailError: string;
  passwordError: string;
  confirmPasswordError: string;
  loginError: string;
  logout: string;
  newName: string;
  wannaChangePassword: string;
  clickHere: string;
  changePassword: string;
  oldPassword: string;
  newPassword: string;
  changePasswordError: string;
  addresses: string;
  addAddress: string;
  street: string;
  houseNumber: string;
  city: string;
  zipCode: string;
  phoneNumber: string;
  floorNumber: string;
  accessCode: string;
  changeAddress: string;
  delete: string;
  updateAddress: string;
  delivery: string;
  pickup: string;
  sortBy: string;
  closeToMe: string;
  grades: string;
  cheapestDelivery: string;
  seeMap: string;
  order: string;
  restaurants: string;
  searchRestaurant: string;
  deliveryClosed: string;
  yourOrder: string;
  basketEmpty: string;
  didntAdd: string;
  notAvailable: string;
  searchMenu: string;
  together: string;
  deliveryCost: string;
  total: string;
  orderNow: string;
  edit: string;
  aboutThisBusiness: string;
  reviews: string;
  info: string;
  overallScore: string;
  outOf: string;
  reviews2: string;
  allReviews: string;
  quality: string;
  deliveryTime: string;
  deliveryHour: string;
  monday: string;
  tuesday: string;
  wednesday: string;
  thursday: string;
  friday: string;
  saturday: string;
  sunday: string;
  fees: string;
  deliveryFee: string;
  feesComment: string;
  paymentMethods: string;
  restaurantInfo: string;
  payU: string;
  creditCard: string;
  cash: string;
  minimumOrder: string;
  chooseIngredients: string;
  addToBasket: string;
  personalData: string;
  additionalInformation: string;
  deliveryAddress: string;
  relevantInformation: string;
  planDelivery: string;
  lackOfAddress: string;
  asSoonAsPossible: string;
  plan: string;
  chooseHour: string;
  selectHour: string;
  confirm: string;
  cancel: string;
  payment: string;
  ordering: string;
  serviceFee: string;
  planPickUp: string;
  selectHourPickUp: string;
  mustLogin: string;
  youAreLogin: string;
  paymentStatus: string;
  paymentNumber: string;
  status: string;
  COMPLETED: string;
  CANCELED: string;
  paymentTitleSuccess: string;
  paymentTitleFailure: string;
  goToOrders: string;
  backToCheckout: string;
  backToMainPage: string;
  orderFrom: string;
  orderNumber: string;
  quantity: string;
  showMenu: string;
  hideMenu: string;
  price: string;
  deliveryMethod: string;
  addOpinion: string;
  addingOpinionFor: string;
  addComment: string;
  locked: string;
  sendCode: string;
  twoFactorCode: string;
  choose: string;
  error2FA: string;
  yourLocation: string;
  closed: string;
  loginWithGoogle: string;
  or: string;
  phoneNumberError: string;
  orderMonitoring: string;
  userData: string;
  toAddress: string;
  userEmail: string,
  notPaid: string;
  paid: string;
  paymentId: string;
  changeStatus: string;
  confirmChangeStatus: string;
  confirmContent: string;
  addRestaurant: string;
  restaurantData: string;
  restaurantName: string;
  restaurantCategory: string;
  restaurantImage: string;
  restaurantLogo: string;
  restaurantAddress: string;
  flatNumber: string;
  restaurantDeliveryData: string;
  deliveryMaxTime: string;
  deliveryMinTime: string;
  deliveryPrice: string;
  minimumPrice: string;
  pickupTime: string;
  restaurantDeliveryHours: string;
  openTime: string;
  closeTime: string;
  restaurantMenu: string;
  addMenu: string;
  category: string;
  image: string;
  ingredients: string;
  addAdditives: string;
  additive: string;
  type: string;
  removeMenuItem: string;
  addMenuItem: string;
  removeAdditive: string;
  availableRestaurants: string;
  editRestaurant: string;
  forgotPassword: string;
  resetPassword: string;
}